+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.4.0] 
 ### Added
 * Logging throughout the download request cycle
 * Logging expiration get/set/check
@@ -394,6 +394,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 2014-08-25
   * working on tests
 
+[1.4.0]: https://github.com/koopjs/koop-agol/compare/v1.3.5..v1.4.0
 [1.3.5]: https://github.com/koopjs/koop-agol/compare/v1.3.4..v1.3.5
 [1.3.4]: https://github.com/koopjs/koop-agol/compare/v1.3.3..v1.3.4
 [1.3.3]: https://github.com/koopjs/koop-agol/compare/v1.3.2..v1.3.3